Changed Auto-Updater to Github

+++ MangaUnhost/Github.cs
@@ -6,61 +6,92 @@ using System.Net;
 using System.Threading;
 using Ionic.Zip;
 
-class AppVeyor {
-    const string UpdateSufix = "-NewAppUpdate.exe";
-    const string Info = "build";
+class GitHub {
 
-    string API = "https://ci.appveyor.com/api/projects/{0}/{1}/";
-    string Artifact = string.Empty;
+    string API = "https://api.github.com/repos/{0}/{1}/releases";
+
+    string cache = null;
+    string Name = null;
     public static string MainExecutable = new Uri(System.Reflection.Assembly.GetExecutingAssembly().CodeBase).LocalPath;
+    public static string TempUpdateDir = Path.GetDirectoryName(MainExecutable) + "\\GitHubRelease\\";
     public static string CurrentVersion {
         get {
             var Version = FileVersionInfo.GetVersionInfo(MainExecutable);
             return Version.FileMajorPart + "." + Version.FileMinorPart;
         }
     }
-    public AppVeyor(string Username, string Project, string Artifact) {
+    public GitHub(string Username, string Project) {
         API = string.Format(API, Username, Project);
-        this.Artifact = Artifact;
 
+        if (!File.Exists(MainExecutable))
+            throw new Exception("Failed to Catch the Executable Path");
+    }
+    public GitHub(string Username, string Project, string Name) {
+        API = string.Format(API, Username, Project);
+        this.Name = Name;
         if (!File.Exists(MainExecutable))
             throw new Exception("Failed to Catch the Executable Path");
     }
 
     public string FinishUpdate() {
-        if (MainExecutable.EndsWith(UpdateSufix)) {
-            string OriginalPath = MainExecutable.Substring(0, MainExecutable.Length - UpdateSufix.Length);
-            Delete(OriginalPath);
-            File.Copy(MainExecutable, OriginalPath);
-            return OriginalPath;
-        } else {
-            if (File.Exists(MainExecutable + UpdateSufix)) {
-                Delete(MainExecutable + UpdateSufix);
-            }
-            string[] OldFiles = Directory.GetFiles(AppDomain.CurrentDomain.BaseDirectory, "*.bak", SearchOption.AllDirectories);
-            foreach (string file in OldFiles)
-                Delete(file);
+        if (FinishUpdatePending()) {
+            int Len = MainExecutable.IndexOf("\\GitHubRelease\\");
 
+            string OriginalPath = MainExecutable.Substring(0, Len);
+            string RunningDir = Path.GetDirectoryName(MainExecutable);
+
+            if (!RunningDir.EndsWith("\\"))
+                RunningDir += '\\';
+            if (!OriginalPath.EndsWith("\\"))
+                OriginalPath += '\\';
+
+            foreach (string File in Directory.GetFiles(RunningDir, "*.*", SearchOption.AllDirectories)) {
+                string Base = File.Substring(RunningDir.Length).TrimStart('\\');
+                string UpPath = RunningDir + Base;
+                string OlPath = OriginalPath + Base;
+
+
+                Delete(OlPath);
+                System.IO.File.Copy(UpPath, OlPath, true);
+            }
+
+            return OriginalPath + Path.GetFileName(MainExecutable);
+        } else {
+            new Thread(() => {
+                int i = 0;
+                while (Directory.Exists(TempUpdateDir) && i < 5) {
+                    try {
+                        Directory.Delete(TempUpdateDir, true);
+                    } catch { Thread.Sleep(1000); i++; }
+                }
+            }).Start();
             return null;
         }
     }
 
     private void Delete(string File) {
         for (int Tries = 0; Tries < 10; Tries++) {
-            Process[] Procs = Process.GetProcessesByName(Path.GetFileNameWithoutExtension(File));
-            foreach (var Proc in Procs) {
-                try {
-                    Proc.Kill();
-                    Thread.Sleep(100);
-                } catch { }
-            }
             try {
+                string ProcName = Path.GetFileNameWithoutExtension(MainExecutable);
+                Process[] Procs = Process.GetProcessesByName(ProcName);
+                int ID = Process.GetCurrentProcess().Id;
+                foreach (var Proc in Procs) {
+                    if (Proc.Id == ID)
+                        continue;
+
+                    try {
+                        Proc.Kill();
+                        Thread.Sleep(100);
+                    } catch { }
+                }
+
                 if (System.IO.File.Exists(File))
                     System.IO.File.Delete(File);
             } catch {
                 Thread.Sleep(100);
                 continue;
             }
+
             break;
         }
     }
@@ -82,41 +113,35 @@ class AppVeyor {
                 return false;//Lst<Curr
             }
             return false;
-        } catch { return false; }
+        } catch (Exception ex){ return false; }
     }
 
+    public bool FinishUpdatePending() {
+        if (MainExecutable.Contains("\\GitHubRelease\\"))
+            return true;
+
+        return false;
+    }
     public void Update() {
         if (!HaveUpdate())
             return;
+
         string Result = FinishUpdate();
         if (Result != null) {
             Process.Start(Result);
             Environment.Exit(0);
         }
 
-        MemoryStream Update = new MemoryStream(Download(API + "artifacts/" +  Artifact.Replace("\\", "/")));
+        MemoryStream Update = new MemoryStream(Download(GetDownloadUrl()));
         var Zip = ZipFile.Read(Update);
-        string TMP = Path.GetTempFileName();
-        if (File.Exists(TMP))
-            File.Delete(TMP);
-        TMP += "\\";
-        if (!Directory.Exists(TMP))
-            Directory.CreateDirectory(TMP);
-        Zip.ExtractAll(TMP, ExtractExistingFileAction.OverwriteSilently);
+        try {
+            if (Directory.Exists(TempUpdateDir))
+                Directory.Delete(TempUpdateDir, true);
+        } catch { }
 
-        foreach (string File in Directory.GetFiles(TMP, "*.*", SearchOption.AllDirectories)) {
-            string Base = File.Substring(TMP.Length, File.Length - TMP.Length);
-            string Output = AppDomain.CurrentDomain.BaseDirectory + Base;
-            if (Output == MainExecutable) {
-                Output += UpdateSufix;
-            }
-
-            Backup(Output);
-            System.IO.File.Move(File, Output);
-        }
-        Directory.Delete(TMP);
-
-        Process.Start(MainExecutable + UpdateSufix);
+        Directory.CreateDirectory(TempUpdateDir);
+        Zip.ExtractAll(TempUpdateDir, ExtractExistingFileAction.OverwriteSilently);
+        Process.Start(TempUpdateDir + Path.GetFileName(MainExecutable));
         Environment.Exit(0);
     }
 
@@ -175,11 +200,38 @@ class AppVeyor {
         }
     }
 
+   
+    
     private string GetLastestVersion() {
-        string Reg = "\"version\":[\\s]*\"([0-9.]*)\"";
-        string XML = new WebClient().DownloadString(API);
-        var a = System.Text.RegularExpressions.Regex.Match(XML, Reg);
+        string Reg = @"\""name\"":[\s]*\""[A-z]*([0-9.]*)[A-z]*\""";
+        var a = System.Text.RegularExpressions.Regex.Match(GetApiResult(), Reg);
         return a.Groups[1].Value;
     }
+    private string GetDownloadUrl() {
+        string Reg = @"\""browser_download_url\"":[\s]*\""(.*)\""";
+        var a = System.Text.RegularExpressions.Regex.Match(GetApiResult(), Reg);
+        if (Name == null)
+            return a.Groups[1].Value;
+        for (int i = 0; i < a.Groups.Count; i++) {
+            string URL = a.Groups[i].Value;
+            URL = URL.Split('?')[0].ToLower();
+            if (URL.EndsWith(Name.ToLower()))
+                return a.Groups[i].Value;
+        }
 
+        throw new FileNotFoundException("Github Release File Not Found.");
+    }
+
+    private string GetApiResult() {
+        if (cache != null)
+            return cache;
+
+        ServicePointManager.SecurityProtocol = (SecurityProtocolType)0x00000FF0;
+        ServicePointManager.ServerCertificateValidationCallback = (sender, cert, chain, sslPolicyErrors) => true;
+
+        WebClient Client = new WebClient();
+        Client.Headers.Add("user-agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/66.0.3359.170 Safari/537.36 OPR/53.0.2907.68");
+        cache = Client.DownloadString(API);
+        return cache;
+    }
 }
